refactor: move original methods into a subdirectory
