+++ src/api/index.js
@@ -1,0 +1,3 @@
+export { getPatches } from './getPatches.js'
+export { getDirectory } from './getDirectory.js'
+export { getFile } from './getFile.js'
