+++ src/index.js
@@ -1,3 +1,1 @@
-export { getPatches } from './getPatches.js'
-export { getDirectory } from './getDirectory.js'
-export { getFile } from './getFile.js'
+export * from './api/index.js'
